ci: add dependabot and ci workflow configuration
Sets up a standard set of CI workflows to automatically build, test, and
pubish the crate (using release-plz). Also adds a dependabot
configuration to automatically update dependencies.

This is mostly a copy of the workflows I use in my own projects, without
the msrv check, as I don't think that's necessary for this crate.

See https://github.com/joshka/github-workflows/ for more information.

With the release-plz workflow, anytime you push to main, it will create
a PR with the new version number and changelog. You can then merge this
PR to publish the new version to crates.io. See
https://github.com/joshka/tui-scrollview/pull/25 for an example of what
this looks like.